Add lcov.info to npmignore.
lcov.info's presence (in the npm package) confuses Drupal 7, which
interprets any .info file inside of a Drupal theme or Drupal module to
be a Drupal sub-theme or Drupal sub-module.

Drupal themes are increasingly using grunt and grunt plugins, and this
npm package is used by at least grunt-grunticon.

This issue manifests itself in Drupal as a menu rebuild failing when
clearing caches.