removeViewBox: Handle more forms
The spec says that each term in viewBox is "separated by whitespace and/or a comma", but the regex here does not allow a comma.

This updated regex allows a comma. It is, however, still more strict that the spec, in that it requires a space (AFTER the comma, if it exists).