Optimize repeated commands
According to the SVG spec, repeated commands can be omitted. While this is not shorter if a space is needed in its place, the space can be omitted if the next number is negative. This optimization does exactly that. The reduction in file size is very small, in my tests much less than 1%, but it does help and it's a relatively small change.