Fix erroneus omission of attribute corrections
The cleanNumericValues plugin contained a condition that causes the
correction of values to be scrapped even though the correction should be
applied. A simply example would be the omission to change the value of
the width attribute, if the value is "20mm".
This condition was removed and the remaining three lines simplified into
two.