Clarify how to enable/configure builtin plugins
I just spent quite a while trying to upgrade to svgo 2.3 because I was thinking that all the configuration I was passing to `extendDefaultPlugins()` previously should now be in `preset-default` overrides. 

Just adding some examples of enabling other builtin plugins could clarify it? 

Maybe I was just being stupid, but a thought :)