docs: fix all mdn urls by specifying en-US (#2182)
MDN used to redirect hits to `/docs` to the users locale, however they
stopped doing that for some reason. This broke all MDN links on our
documentation/JSDocs.

This inserts `en-US` to all MDN URLs which fixes the links.
(Unfortunately, this means all users may be taken to the English site
regardless of their language preference.)