perf: optimiztions to reduce regression test runtime (#2135)
Small group of changes that at reduce the regression test runtime
by about 14%.

Motivation for this change in particular were the large PNGs usually
allocated for even the smallest SVGs in the test data which both increases
the time spent allocating as well as the time pixelmatch would take for the
comparison.

The other changes improve not only the test runtime but also the normal
SVG processing given the input data has cases caught by
removeHiddenElems which also has the by far most significant of the 3
changes. All the changes done for baseline performance optimization are
very unlikely to cause any harm and would need specific sets of input
SVGs mostly consisting of visibility checks regarding the element style
to impact performance in a negative way. In my local tests these changes
gave an additional ~7-8% time reduction with the svgo data set.